plugin_example/vera_sql_query_assistant/pyproject.toml
[project]
name = "vera-sql-query-assistant"
version = "1.0.0"
description = "Example Vera implementation that provides tests for a SQL Query Assistant feature"
readme = "README.md"
authors = [
    { name = "Tal Shapir", email = "talshapir@google.com" },
]
requires-python = ">=3.14"
license-files = ["LICENSE"]
license = "Apache-2.0"
dependencies = [
    "anyio>=4.12.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "vera>=1.0.0",
]
keywords = [
    "vera",
    "example",
    "plugin",
]
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
]

[project.urls]
Repository = "https://github.com/google/vera"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
    "ty>=0.0.10",
]

[project.entry-points.vera]
vera_sql_query_assistant = "vera_sql_query_assistant.plugin_impl"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
pythonpath = "src"
addopts = [
    "--import-mode=importlib",
]
testpaths = [
    "tests/",
]

[tool.uv.sources]
vera = { path = "../../vera" }

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
